+++ docs/conf.py
@@ -290,6 +290,7 @@ texinfo_documents = [
 # intersphinx
 intersphinx_mapping = {
     'python': ('http://docs.python.org/3', None),
+    'psycopg2': ('http://initd.org/psycopg/docs', None),
 }
 
 autodoc_member_order = 'bysource'