+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## [Unreleased][unreleased]
 ### Added
 - `parse_acl` function to call `parse_acl_item` on each item in a list.
+- Documentation on `parse_acl_item` about ensuring `subname` is a valid
+  identifier if it is to be executed.
+
 ### Changed
 - Renamed singular functions like `get_table_acls` to `get_table_acl`.
 - `get_all_table_acls` and `get_table_acl` also return views, materialized